+++ cver/units/sine.h
@@ -12,7 +12,7 @@ typedef struct Sine_struct {
     Unit unit;
     IO* output;
     IO* freq_in;
-    double phase;
+    float phase;
 } Sine;
 
 Module* Sine_new();
